+++ Example/FWFramework/Classes/Module/Test/Index/TestController.swift
@@ -27,9 +27,10 @@ class TestController: UIViewController {
             ["Codable", "TestCodableController"],
             ["Request", "TestRequestController"],
             ["Socket", "TestSocketController"],
+            ["Purchase", "TestPurchaseController"],
             ["AudioPlayer", "TestAudioController"],
             ["VideoPlayer", "TestVideoController"],
-            ["AudioRecorder", "TestRecorderController"],
+            ["AudioRecorder", "TestRecorderController"]
         ]],
         ["Toolkit", [
             ["Adaptive", "TestAdaptiveController"],