+++ Sources/FWFramework/Plugin/Image/ImageCoder.swift
@@ -1,5 +1,5 @@
 //
-//  AnimatedImage.swift
+//  ImageCoder.swift
 //  FWFramework
 //
 //  Created by wuyong on 2022/8/23.