+++ Sources/FWUIKit/Plugin/Picker/ImagePickerControllerImpl.swift
@@ -1,5 +1,5 @@
 //
-//  ImagePickerPluginImpl.swift
+//  ImagePickerControllerImpl.swift
 //  FWFramework
 //
 //  Created by wuyong on 2022/8/23.